requirements.txt
Django
Django-phonenumber-field
virtual-env
